Fix examples
