+++ examples/demo.rs
@@ -4,8 +4,9 @@ use embedded_graphics_simulator::{
     OutputSettingsBuilder, SimulatorDisplay, SimulatorEvent, Window,
 };
 use lvgl;
+use lvgl::style::Style;
 use lvgl::widgets::{Label, LabelAlign};
-use lvgl::{Align, Color, Object, Style, UI};
+use lvgl::{Align, Color, ObjPart, Object, State, UI};
 use lvgl_sys;
 use std::sync::{mpsc, Arc, Mutex};
 use std::thread::sleep;
@@ -29,30 +30,29 @@ fn main() -> Result<(), String> {
     // Create screen and widgets
     let mut screen = ui.scr_act();
 
-    let font_roboto_28 = unsafe { &lvgl_sys::lv_font_roboto_28 };
+    let font_roboto_28 = unsafe { &lvgl_sys::lv_theme_get_font_normal() };
     let font_noto_sans_numeric_28 = unsafe { &noto_sans_numeric_80 };
 
-    let mut screen_style = Default::default();
-    screen_style.set_body_main_color(Color::from_rgb((0, 0, 0)));
-    screen_style.set_body_grad_color(Color::from_rgb((0, 0, 0)));
-    screen_style.set_body_radius(0);
-    screen.set_style(screen_style);
+    let mut screen_style = Style::default();
+    screen_style.set_bg_color(State::DEFAULT, Color::from_rgb((0, 0, 0)));
+    screen_style.set_radius(State::DEFAULT, 0);
+    screen.add_style(ObjPart::Main, screen_style);
 
     let mut time = Label::new(&mut screen);
-    let mut style_time = Default::default();
-    style_time.set_text_font(font_noto_sans_numeric_28);
-    style_time.set_text_color(lvgl::Color::from_rgb((255, 255, 255)));
-    time.set_style(style_time);
-    time.set_align(&mut screen, Align::InLeftMid, 20, 0);
+    let mut style_time = Style::default();
+    //style_time.set_text_font(font_noto_sans_numeric_28);
+    style_time.set_text_color(State::DEFAULT, Color::from_rgb((255, 255, 255)));
+    time.add_style(ObjPart::Main, style_time);
+    time.set_align(&mut screen, Align::Center, 0, 0);
     time.set_text("20:46");
     time.set_width(240);
     time.set_height(240);
 
     let mut bt = Label::new(&mut screen);
-    let mut style_bt = Default::default();
-    style_bt.set_text_font(font_roboto_28);
+    let mut style_bt = Style::default();
+    //style_bt.set_text_font(font_roboto_28);
     let style_power = style_bt.clone();
-    bt.set_style(style_bt);
+    bt.add_style(ObjPart::Main, style_bt);
     bt.set_width(50);
     bt.set_height(80);
     bt.set_recolor(true);
@@ -61,7 +61,7 @@ fn main() -> Result<(), String> {
     bt.set_align(&mut screen, Align::InTopLeft, 0, 0);
 
     let mut power = Label::new(&mut screen);
-    power.set_style(style_power);
+    power.add_style(ObjPart::Main, style_power);
     power.set_recolor(true);
     power.set_width(80);
     power.set_height(20);
